+++ webflow/full/bin/live-reload.js
@@ -1,0 +1,1 @@
+new EventSource(`${SERVE_ORIGIN}/esbuild`).addEventListener('change', () => location.reload());