+++ changelog.md
@@ -42,3 +42,7 @@
 # 17/03/23
 
 - [x] VITE → added minimal tween library
+
+# 04/04/23
+
+- [x] WEBFLOW → added finsweet based esbuild script